Rename meshlet module
